refactor: extract quickpay coordinator
